Refactor how metadata are discoverd
Instead of listing all the AYS ID in the config file. We specify a directory where the flist files are located and we load them after sorting.
Metadata are loaded in Descending order. So zz.flist is the first aa.flist is the last.

This allow to create some priority between flist. It's usefull for debuging.
If you want to test a new AYS, just push a 00.flist in the directory, trigger metdata reload.

Don't purge binary file at reload or startup, just remove metdata tree instead